+++ examples/data-sources/linkridgecloud_account_services/data-source.tf
@@ -1,0 +1,4 @@
+data "linkridgecloud_account_services" "planned" {
+  account_id = "acct_local_qr_demo"
+  status     = "planned"
+}